Renamed Map.js and change some links to Links from react router dom
